fix : gameId 타입 변경

+++ backend/game-service/src/main/java/yare/yare/domain/game/controller/GameController.java
@@ -42,7 +42,7 @@ public class GameController {
 
     @GetMapping("/{gameId}/details")
     public ResponseDto<GameDetailsRes> gameDetails(
-            @PathVariable Integer gameId) {
+            @PathVariable Long gameId) {
 
         GameDetailsRes result = gameService.findGame(gameId);
